+++ src/drem/extract/download.py
@@ -1,9 +1,11 @@
-#!/usr/bin/env python
-
 from pathlib import Path
+from typing import Any
+from typing import Optional
 
 import requests
 
+from prefect import Task
+from prefect.utilities.tasks import defaults_from_attrs
 from tqdm import tqdm
 
 
@@ -41,3 +43,66 @@ def download(url: str, filepath: Path) -> None:
 
         response.raise_for_status()
         download_file_from_response(response, filepath)
+
+
+def _raise_for_empty_arguments(*args):
+
+    empty_args = [arg for arg in args if arg is None]
+
+    if any(empty_args) is None:
+        raise ValueError(f"At least one of {args} is None.")
+
+
+class Download(Task):
+    """Download file directly from URL.
+
+    Example:
+    Save `data.csv` to current working directory:
+    ```python
+    download_data = Download(
+        url=http://www.urltodata.ie/data.csv, savepath='data.csv',
+    )
+    download_data.run()
+    ```
+
+    Args:
+        Task (Task): see https://docs.prefect.io/core/concepts/tasks.html
+    """
+
+    def __init__(
+        self, url: Optional[str] = None, savepath: Optional[Path] = None, **kwargs: Any,
+    ):
+        """Initialise 'Download' Task.
+
+        Args:
+            url (Optional[str], optional): Direct URL link to data such as
+                http://www.urltodata.ie/data.csv. Defaults to None.
+            savepath (Optional[Path], optional): Data savepath. Defaults to None.
+            **kwargs (Any): see  https://docs.prefect.io/core/concepts/tasks.html
+        """
+        self.url = url
+        self.savepath = savepath
+
+        super().__init__(**kwargs)
+
+    @defaults_from_attrs("url", "savepath")
+    def run(self, url: Optional[str] = None, savepath: Optional[Path] = None) -> None:
+        """Download data directly from URL to savepath.
+
+        Args:
+            url (Optional[str], optional): Direct URL link to data such as
+                http://www.urltodata.ie/data.csv. Defaults to None.
+            savepath (Optional[Path], optional): Data savepath. Defaults to None.
+        """
+        if savepath is None:
+            savepath = self.savepath
+
+        if url is None:
+            url = self.url
+
+        _raise_for_empty_arguments(url, savepath)
+
+        with requests.get(url=url, stream=True) as response:
+
+            response.raise_for_status()
+            download_file_from_response(response, savepath)